move schema files into lib/schemas and update references
