+++ lib/schemas/events-schema.js
@@ -1,4 +1,4 @@
-const languages = require('./languages')
+const languages = require('../languages')
 
 const context = {
   type: 'object',
