+++ lib/schemas/products-schema.js
