+++ lib/schemas/site-tree-schema.js
