+++ tests/unit/versions.js
@@ -1,7 +1,7 @@
 const revalidator = require('revalidator')
 const allVersions = require('../../lib/all-versions')
 const { latest } = require('../../lib/enterprise-server-releases')
-const schema = require('../../lib/versions-schema')
+const schema = require('../../lib/schemas/versions-schema')
 const { getJSON } = require('../helpers/supertest')
 const nonEnterpriseDefaultVersion = require('../../lib/non-enterprise-default-version')
